# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1",
    "title": "sysfind",
    "description": "sysfind"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "127.0.0.1"
    }
  ],
  "tags": [
    {
      "name": "Tag",
      "description": "Tag"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/operatingSystem/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "processorArchitecture": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/operatingSystem/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/operatingSystem/family/{family}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "family",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {},
    "securitySchemas": {
      "cookie": {
        "type": "apiKey",
        "in": "header",
        "name": "LOGIN-COOKIE"
      }
    }
  }
}